examples/dynamics/requirements.txt
boto3
python-dotenv
msal